refactor code  with js module
