+++ resources/js/constants.js
@@ -2,3 +2,5 @@ const ITEM_TYPE = {
   FILE: "1",
   FOLDER: "2",
 };
+
+export { ITEM_TYPE };
